Fix: GroupRepository에서 투표 로직 분리
